Swapped the internal ints in Rectangle and Size for floats.
Did this to increase presicion when rendering with OpenGL.

Still having some VSYNC problems.

+++ src/core/GLRenderableObject.java
@@ -21,7 +21,7 @@ public interface GLRenderableObject {
 	 *            The canvas to draw to.
 	 */
 	public void render(Rectangle object, Rectangle target,
-			GLAutoDrawable canvas, int zIndex);
+			GLAutoDrawable canvas, float zIndex);
 
 	/**
 	 * Updates the state of the GLRenderableObject.
